python_backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
google-generativeai
httpx
websockets
python-multipart # For FastAPI file uploads
